Rename HealthChecksUI sample
