+++ samples/HealthChecksUI/HealthChecksUI.ApiService/Program.cs
