+++ samples/HealthChecksUI/HealthChecksUI.ApiService/appsettings.json
