+++ samples/HealthChecksUI/HealthChecksUI.AppHost/Program.cs
@@ -1,0 +1,16 @@
+﻿var builder = DistributedApplication.CreateBuilder(args);
+
+var cache = builder.AddRedis("cache");
+
+var apiService = builder.AddProject<Projects.HealthChecksUI_ApiService>("apiservice")
+    .WithHttpEndpoint(name: "healthchecks");
+
+var webFrontend = builder.AddProject<Projects.HealthChecksUI_Web>("webfrontend")
+    .WithReference(cache)
+    .WithReference(apiService);
+
+builder.AddHealthChecksUI("healthchecksui")
+    .WithReference(apiService)
+    .WithReference(webFrontend, endpointName: "http");
+
+builder.Build().Run();
