+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Layout/MainLayout.razor
