+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Layout/NavMenu.razor
@@ -1,6 +1,6 @@
 ﻿<div class="top-row ps-3 navbar navbar-dark">
     <div class="container-fluid">
-        <a class="navbar-brand" href="">AspireApp33</a>
+        <a class="navbar-brand" href="">HealthChecksUI</a>
     </div>
 </div>
