+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Layout/NavMenu.razor.css
