+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Pages/Counter.razor
