+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Pages/Home.razor
