+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Pages/Weather.razor
