+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/Routes.razor
