+++ samples/HealthChecksUI/HealthChecksUI.Web/Components/_Imports.razor
@@ -7,5 +7,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.OutputCaching
 @using Microsoft.JSInterop
-@using AspireApp33.Web
-@using AspireApp33.Web.Components
+@using HealthChecksUI.Web
+@using HealthChecksUI.Web.Components
