+++ samples/HealthChecksUI/HealthChecksUI.Web/appsettings.json
