+++ samples/HealthChecksUI/HealthChecksUI.Web/wwwroot/favicon.png
